% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2calc_genoprob.R
\name{feather2calc_genoprob}
\alias{feather2calc_genoprob}
\title{Store genotype probabilities in feather database}
\usage{
feather2calc_genoprob(object)
}
\arguments{
\item{object}{Object of class \code{\link{feather_genoprob}}.}
}
\value{
An object of class \code{\link[qtl2]{calc_genoprob}}.
}
\description{
Uses package feather to convert R object created in R/qtl2 for fast access.
}
\details{
The genotype probabilities are extracted from 1-2 feather databases. Each chromosome is extracted in turn.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
dir <- tempdir()
fprobs <- feather_genoprob(probs, "grav2", dir)
nprobs <- feather2calc_genoprob(fprobs)

}
\keyword{utilities}
