% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather2fst_genoprob.R
\name{feather2fst_genoprob}
\alias{feather2fst_genoprob}
\title{Migrate feather_genoprob to fst_genoprob database.}
\usage{
feather2fst_genoprob(fprobs, verbose = TRUE)
}
\arguments{
\item{verbose}{Show warning of feather creation if \code{TRUE} (default).}

\item{fprob}{object of class \code{feather_genoprob}.}

\item{fdir}{Directory for feather database.}
}
\value{
An object of class \code{fst_genoprob}.
}
\description{
See package \code{R/qtl2fst} for faster database version.
}
\details{
The genotype probabilities are stored using package \code{fst} rather package \code{feather} for faster access.
If the feather directory address is relative, then be sure to use \code{\link{setwd}} to get to the parent directory.
}
\keyword{utilities}
