% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_genoprob.R
\name{feather_genoprob}
\alias{feather_genoprob}
\title{Store genotype probabilities in feather database}
\usage{
feather_genoprob(genoprob, fbase, fdir = ".", verbose = TRUE)
}
\arguments{
\item{genoprob}{Object of class \code{"calc_genoprob"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}
and \code{\link[qtl2]{calc_genoprob}}.}

\item{fbase}{Base of filename for feather database.}

\item{fdir}{Directory for feather database.}

\item{verbose}{Show warning of feather creation if \code{TRUE} (default).}
}
\value{
A list containing the attributes of \code{genoprob}
and the address for the created feather database.
Components are:
\itemize{
\item \code{dim} - List of all dimensions of 3-D arrays.
\item \code{dimnames} - List of all dimension names of 3-D arrays.
\item \code{is_x_chr} - Vector of all is_x_chr attributes.
\item \code{chr} - Vector of (subset of) chromosome names for this object.
\item \code{ind} - Vector of (subset of) individual names for this object.
\item \code{mar} - Vector of (subset of) marker names for this object.
\item \code{feather} - Names of feather databases: \code{A} = autosome database, \code{X} = X chromosome database (if needed).
}
}
\description{
Uses package feather to convert R object created in R/qtl2 for fast access.
}
\details{
The genotype probabilities are stored in separate databases for each chromosome
as tables of (indivduals*genotypes) x (positions) in directory \code{feather}. 
The \code{dim}, \code{dimnames} and \code{is_x_chr} elements of the object
have information about the entire feather database.
If a \code{feather_genoprob} object is a subset of another such object, 
the \code{chr}, \code{ind}, and \code{mar} contain information about what is in the subset.
However, the \code{feather} databases are not altered in a subset, and can be restored by
\code{\link{feather_genoprob_restore}}. The actual elements of a \code{feather_genoprob}
object are only accessible to the user after a call to \code{\link[base]{unclass}}; instead
the usual access to elements of the object invoke \code{\link{subset.feather_genoprob}}.
}
\examples{
library(qtl2)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
dir <- tempdir()
fprobs <- feather_genoprob(probs, "grav2", dir)
}
\keyword{utilities}
