% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_genoprob_list.R
\name{feather_genoprob_list}
\alias{feather_genoprob_list}
\title{Extract list element(s) from feather_genoprob object}
\usage{
feather_genoprob_list(object, element, ...)
}
\arguments{
\item{object}{Object of class \code{\link{feather_genoprob}}.}

\item{element}{Element to extract}

\item{...}{Additional elements. More than one causes return of list.}
}
\value{
Element or list of elements from \code{\link{feather_genoprob}} object.
}
\description{
Interprets object as an unclassed list to get components.
}
\details{
Object is unclassed and elements are extracted. All elements are extracted if only object supplied.
See \code{\link{feather_genoprob}} return description for details.
}
\examples{
library(qtl2geno)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
map <- insert_pseudomarkers(grav2$gmap, step=1)
probs <- calc_genoprob(grav2, map, error_prob=0.002)
fprobs <- feather_genoprob(probs, "my.feather")
feather_genoprob_list(fprobs, "feather")
sapply(feather_genoprob_list(fprobs, "chr", "ind", "mar"), length)

}
\keyword{utilities}
