% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feather_genoprob.R
\name{feather_genoprob}
\alias{feather_genoprob}
\title{Store genotype probabilities in feather database}
\usage{
feather_genoprob(genoprob, basename, dirname = ".")
}
\arguments{
\item{genoprob}{Object of class \code{"calc_genoprob"}. For details, see the
\href{http://kbroman.org/qtl2/assets/vignettes/developer_guide.html}{R/qtl2 developer guide}
and \code{\link[qtl2geno]{calc_genoprob}}.}

\item{basename}{Base of fileame for feather database.}

\item{dirname}{Directory for feather database.}
}
\value{
A list containing the attributes of \code{genoprob} and the address for the created feather database.
}
\description{
Uses package feather to convert R object created in R/qtl2geno for fast access.
}
\details{
The genotype probabilities are stored in 1-2 databases
as a table of (indivduals*genotypes) x (positions). The first database has all autosome positions, 
identified by marker or pseudomarker name. The optional second database is for the X chromosome if present.
}
\examples{
library(qtl2geno)
grav2 <- read_cross2(system.file("extdata", "grav2.zip", package="qtl2geno"))
probs <- calc_genoprob(grav2, step=1, error_prob=0.002)
fprobs <- feather_genoprob(probs, "my.feather")
}
\keyword{utilities}
